# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admins": {
      "post": {
        "operationId": "AdminController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAdminDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid input."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "409": {
            "description": "Email already exists."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new admin user and profile",
        "tags": [
          "Admins Management"
        ]
      },
      "get": {
        "operationId": "AdminController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of admin profiles.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminProfileResponseDto"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a list of all admin profiles",
        "tags": [
          "Admins Management"
        ]
      }
    },
    "/admins/profile": {
      "get": {
        "operationId": "AdminController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Detail profil admin, termasuk perusahaan dan API keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mendapatkan profil admin yang sedang login",
        "tags": [
          "Admins Management"
        ]
      }
    },
    "/admins/{id}": {
      "get": {
        "operationId": "AdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin Profile ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin profile details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Admin profile not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific admin profile by ID",
        "tags": [
          "Admins Management"
        ]
      },
      "patch": {
        "operationId": "AdminController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin Profile ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAdminDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin profile updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or invalid input."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Admin profile not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update an admin profile",
        "tags": [
          "Admins Management"
        ]
      },
      "delete": {
        "operationId": "AdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Admin Profile ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Admin deleted successfully."
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Admin profile not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete an admin profile (and associated user)",
        "tags": [
          "Admins Management"
        ]
      }
    },
    "/admins/{adminId}/roles/{roleId}": {
      "post": {
        "operationId": "AdminController_assignRole",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "description": "Admin Profile ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role assigned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Admin or Role not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a role to an admin",
        "tags": [
          "Admins Management"
        ]
      },
      "delete": {
        "operationId": "AdminController_removeRole",
        "parameters": [
          {
            "name": "adminId",
            "required": true,
            "in": "path",
            "description": "Admin Profile ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "roleId",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role removed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProfileResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden. Missing required permissions."
          },
          "404": {
            "description": "Admin or Role not found."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Remove a role from an admin",
        "tags": [
          "Admins Management"
        ]
      }
    },
...
  },
  "info": {
    "title": "SSO Boilerplate API",
    "description": "API documentation for the SSO Boilerplate application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API Key untuk autentikasi akses API"
      }
    },
    "schemas": {
...
      "RoleResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "admin"
          },
          "description": {
            "type": "object",
            "example": "Administrator role",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "ApiKeyEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID unik API key",
            "example": 1
          },
          "key": {
            "type": "string",
            "description": "API Key string (hanya ditampilkan saat pembuatan)",
            "example": "7c8a5d69f43e5c8a94d7ec68"
          },
          "name": {
            "type": "string",
            "description": "Nama untuk identifikasi API key",
            "example": "Production Key"
          },
          "description": {
            "type": "string",
            "description": "Deskripsi tentang tujuan penggunaan API key",
            "example": "API key untuk akses ke production environment"
          },
          "isActive": {
            "type": "boolean",
            "description": "Status API key (aktif/tidak aktif)",
            "example": true
          },
          "companyId": {
            "type": "number",
            "description": "ID perusahaan yang memiliki API key",
            "example": 1
          },
          "lastUsedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Terakhir kali API key digunakan",
            "example": "2023-05-20T15:45:30Z"
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal kadaluarsa API key (jika ada)",
            "example": "2023-12-31T23:59:59Z"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal pembuatan API key",
            "example": "2023-01-15T10:30:45Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Tanggal update terakhir API key",
            "example": "2023-01-15T10:30:45Z"
          }
        },
        "required": [
          "id",
          "key",
          "name",
          "isActive",
          "companyId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CompanyResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The unique identifier of the company"
          },
          "name": {
            "type": "string",
            "example": "Acme Corp",
            "description": "The name of the company"
          },
          "description": {
            "type": "object",
            "example": "Leading provider of innovative solutions",
            "description": "Company description",
            "nullable": true
          },
          "apiKeys": {
            "description": "API keys belonging to the company",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyEntity"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateAdminDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "Email address for the admin user"
          },
          "password": {
            "type": "string",
            "example": "AdminPass123",
            "description": "Password with minimum 8 characters",
            "minLength": 8
          },
          "name": {
            "type": "string",
            "example": "John Admin",
            "description": "Full name of the admin",
            "maxLength": 150
          },
          "companyId": {
            "type": "number",
            "example": 1,
            "description": "ID of the company this admin belongs to (optional)"
          },
          "roleIds": {
            "example": [
              1,
              2
            ],
            "description": "Array of role IDs assigned to this admin",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "email",
          "password",
          "name",
          "roleIds"
        ]
      },
      "AdminProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "The unique identifier of the admin profile."
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "The name of the admin."
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "The email of the admin user."
          },
          "title": {
            "type": "object",
            "example": "Administrator",
            "description": "Optional title or position.",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "example": "+1234567890",
            "description": "Optional contact phone number.",
            "nullable": true
          },
          "roles": {
            "description": "Roles assigned to the admin.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoleResponseDto"
            }
          },
          "company": {
            "description": "Company associated with the admin.",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanyResponseDto"
              }
            ]
          },
          "apiKeys": {
            "description": "API keys for the admin's company",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiKeyEntity"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateAdminDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe Updated",
            "description": "Updated name for the admin",
            "maxLength": 150
          },
          "companyId": {
            "type": "object",
            "example": 2,
            "description": "Updated company ID (set null to remove company association)",
            "nullable": true
          },
          "roleIds": {
            "example": [
              1,
              3,
              5
            ],
            "description": "Updated array of role IDs",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
...
    }
  }
}